+++ include/aikido/statespace/CartesianProduct.hpp
@@ -35,6 +35,9 @@ public:
   /// \return new \c ScopedState
   ScopedState createState() const;
 
+  /// Creates an identical clone of \c stateIn.
+  ScopedState cloneState(const StateSpace::State* stateIn) const;
+
   /// Gets number of subspaces.
   ///
   /// \return number of subspaces