+++ tests/statespace/test_CartesianProduct.cpp
@@ -1,3 +1,4 @@
+#include <dart/math/Helpers.hpp>
 #include <gtest/gtest.h>
 #include <aikido/statespace/CartesianProduct.hpp>
 #include <aikido/statespace/Rn.hpp>
@@ -11,6 +12,28 @@ using aikido::statespace::R3;
 using aikido::statespace::SO2;
 using aikido::statespace::SO3;
 using aikido::statespace::SE2;
+
+TEST(CartesianProduct, Clone)
+{
+  CartesianProduct space({std::make_shared<SO2>(), std::make_shared<R2>()});
+
+  for (auto i = 0u; i < 5u; ++i)
+  {
+    auto s1 = space.createState();
+    const auto angle = dart::math::random(-M_PI, M_PI);
+    s1.getSubStateHandle<SO2>(0).fromAngle(angle);
+    s1.getSubStateHandle<R2>(1).setValue(Eigen::Vector2d::Random());
+
+    auto s2 = s1.clone();
+
+    EXPECT_DOUBLE_EQ(
+        s1.getSubStateHandle<SO2>(0).toAngle(),
+        s2.getSubStateHandle<SO2>(0).toAngle());
+    EXPECT_TRUE(
+        s1.getSubStateHandle<R2>(1).getValue().isApprox(
+            s2.getSubStateHandle<R2>(1).getValue()));
+  }
+}
 
 TEST(CartesianProduct, Compose)
 {