+++ tests/statespace/test_SE2.cpp
@@ -1,7 +1,28 @@
+#include <dart/math/Helpers.hpp>
 #include <gtest/gtest.h>
 #include <aikido/statespace/SE2.hpp>
 
 using aikido::statespace::SE2;
+
+TEST(SE2, Clone)
+{
+  SE2 se2;
+
+  for (auto i = 0u; i < 5u; ++i)
+  {
+    const auto angle = dart::math::random(-M_PI, M_PI);
+    Eigen::Isometry2d pose = Eigen::Isometry2d::Identity();
+    pose.rotate(Eigen::Rotation2Dd(angle));
+    pose.translation() = Eigen::Vector2d::Random();
+
+    auto s1 = se2.createState();
+    s1.setIsometry(pose);
+
+    auto s2 = s1.clone();
+
+    EXPECT_TRUE(s1.getIsometry().isApprox(s2.getIsometry()));
+  }
+}
 
 TEST(SE2, Compose)
 {